* application-src/amgtar.c, server-src/holding.c: Cast getpid() to int   before printing it.
git-svn-id: https://svn.code.sf.net/p/amanda/code/amanda/trunk@5408 a8d146d6-cc15-0410-8900-af154a0219e0

+++ application-src/amgtar.c
@@ -1200,7 +1200,7 @@ amgtar_restore(
 	if (argument->dle.disk) {
 	    sdisk = sanitise_filename(argument->dle.disk);
 	} else {
-	    sdisk = g_strdup_printf("no_dle-%d", getpid());
+	    sdisk = g_strdup_printf("no_dle-%d", (int)getpid());
 	}
 	exclude_filename= g_strjoin(NULL, AMANDA_TMPDIR, "/", "exclude-", sdisk,  NULL);
 	exclude_list = fopen(argument->dle.exclude_list->first->name, "r");
@@ -1255,7 +1255,7 @@ amgtar_restore(
 	if (argument->dle.disk) {
 	    sdisk = sanitise_filename(argument->dle.disk);
 	} else {
-	    sdisk = g_strdup_printf("no_dle-%d", getpid());
+	    sdisk = g_strdup_printf("no_dle-%d", (int)getpid());
 	}
 	include_filename = g_strjoin(NULL, AMANDA_TMPDIR, "/", "include-", sdisk,  NULL);
 	include = fopen(include_filename, "w");
